Restructure API resources
