+++ Jellyfin.Plugin.Listenbrainz/Models/Listenbrainz/Requests/ValidateTokenRequest.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using static Jellyfin.Plugin.Listenbrainz.Resources.Listenbrainz;
+using Jellyfin.Plugin.Listenbrainz.Resources.Listenbrainz;
 
 namespace Jellyfin.Plugin.Listenbrainz.Models.Listenbrainz.Requests
 {
