+++ Jellyfin.Plugin.Listenbrainz/Resources/Listenbrainz/Api.cs
@@ -1,0 +1,18 @@
+namespace Jellyfin.Plugin.Listenbrainz.Resources.Listenbrainz
+{
+    /// <summary>
+    /// Listenbrainz API resources.
+    /// </summary>
+    public static class Api
+    {
+        /// <summary>
+        /// API version.
+        /// </summary>
+        public const string Version = "1";
+
+        /// <summary>
+        /// API base URL.
+        /// </summary>
+        public const string BaseUrl = "api.listenbrainz.org";
+    }
+}
