+++ Jellyfin.Plugin.Listenbrainz/Resources/Musicbrainz/Api.cs
@@ -1,0 +1,13 @@
+namespace Jellyfin.Plugin.Listenbrainz.Resources.Musicbrainz
+{
+    /// <summary>
+    /// Musicbrainz API resources.
+    /// </summary>
+    public static class Api
+    {
+        /// <summary>
+        /// API base URL.
+        /// </summary>
+        public const string BaseUrl = "musicbrainz.org";
+    }
+}
