+++ Jellyfin.Plugin.Listenbrainz/Resources/Musicbrainz/Endpoints.cs
@@ -1,0 +1,13 @@
+namespace Jellyfin.Plugin.Listenbrainz.Resources.Musicbrainz
+{
+    /// <summary>
+    /// Basic API endpoints.
+    /// </summary>
+    public static class Endpoints
+    {
+        /// <summary>
+        /// Recording endpoint.
+        /// </summary>
+        public const string Recording = "recording";
+    }
+}
